refactor: Move Auth Code to Folder
